add tool

+++ data_structure.py
@@ -12,13 +12,19 @@ class LargeDict(dict):
         self.id = str(uuid.uuid4())
         self.in_memory_key_number = 128
         self.all_keys = set()
+        self.permanent = False
 
     def set_storage_dir(self, storage_dir):
         self.storage_dir = storage_dir
+        os.makedirs(os.path.join(self.storage_dir, self.id))
 
     def set_in_memory_key_number(self, num):
         self.in_memory_key_number = num
-        os.makedirs(os.path.join(self.storage_dir, self.id))
+
+    def save(self):
+        self.permanent = True
+        for k in list(super().keys()):
+            self.__save_item(k)
 
     def __getitem__(self, key):
         if key in self.all_keys:
@@ -39,6 +45,9 @@ class LargeDict(dict):
         return super().__delitem__(key)
 
     def __del__(self):
+        if self.permanent:
+            self.save()
+            return
         if self.storage_dir is not None:
             shutil.rmtree(os.path.join(self.storage_dir, self.id))
